reactor nth reducible
